+++ fixtures/ios_app_with_watchapp2/Derived/Sources/Bundle+WatchAppExtension.swift
@@ -1,0 +1,23 @@
+// swiftlint:disable all
+import Foundation
+
+// MARK: - Swift Bundle Accessor
+
+private class BundleFinder {}
+
+extension Foundation.Bundle {
+    /// Since WatchAppExtension is a watch 2 extension, the bundle containing the resources is copied into the final product.
+    static var module: Bundle = {
+        return Bundle(for: BundleFinder.self)
+    }()
+}
+
+// MARK: - Objective-C Bundle Accessor
+
+@objc
+public class WatchAppExtensionResources: NSObject {
+   @objc public class var bundle: Bundle {
+         return .module
+   }
+}
+// swiftlint:enable all